Don't retweet retweets. Too much pain from Twitter API.
We still archive them.

Twitter refuses to retweet a retweet, and while we could extract the
original and try to retweet that, doing that more than once fails too.
The errors returned make it seem like it's because the author account
was protected, but that's a red herring.

We could handle the errors, but as they are hard to categorize, they
tend to leave *.json.fail junk around.

Hopefully we'll see the original tweet in the stream, and can retweet
it once at that time.

TODO: Twitter stream duplication may cause us to try to retweet
multiple times too. That's painful to diagnose unless we keep a
memory and avoid resending requests. I wish it was idempotent :(